big changes; results via map view, display changes
Rather than using a defined radius, we can calculate the radius of the
current map view in meters, which we then pass to Mongo. As we move
around the map, we send requests out to /search which will just get the
locations nearby. Clicking on a location can also re-center, which will
then also bring a new population around this new frame.

Display changes:
 * moving towards em units in CSS
 * Google map + its controls no longer hidden by leftContainer
 * removed all Google map controls except zoom buttons