fix #27; make searching by map explicit
There's now a map control to make this explicit. It's a much smoother
experience, I think. We can make the control prettier, perhaps, but I do
think the words should stay in at least some form.

Searching also will kill the current selection. If we do not do this,
the map would then recenter on the last selected pantry, which I felt
was a bit confusing and strange.